# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_IUser.id-or-email-or-name-or-created_at-or-updated_at_": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IGetUserRes": {
				"$ref": "#/components/schemas/Pick_IUser.id-or-email-or-name-or-created_at-or-updated_at_"
			}
		},
		"securitySchemes": {
			"basic": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			},
			"google_oauth": {
				"type": "oauth2",
				"description": "Google OAuth 2.0",
				"flows": {
					"authorizationCode": {
						"tokenUrl": "https://oauth2.googleapis.com/token",
						"authorizationUrl": "https://accounts.google.com/o/oauth2/auth",
						"scopes": {
							"https://www.googleapis.com/auth/userinfo.email": "Access user email address",
							"https://www.googleapis.com/auth/userinfo.profile": "Access user profile info",
							"https://www.googleapis.com/auth/contacts.readonly": "Access contacts"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "node-express-tsoa-swagger-typeorm-starter",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "WebEXP0528 ",
			"email": "web.dev0528@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/oauth": {
			"get": {
				"operationId": "GetAuthURI",
				"responses": {
					"302": {
						"description": "Redirect"
					}
				},
				"description": "Go to OAuth Page",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/callback": {
			"get": {
				"operationId": "AuthCallback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/info": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGetUserRes"
								}
							}
						}
					}
				},
				"description": "Returns user info",
				"security": [
					{
						"basic": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}